requirements.txt
discord
python-dotenv
google-genai
